chore: fix the machine-applicable half of the clippy backlog
`cargo clippy --workspace --all-targets --all-features` reported 344
unique warnings on trunk (deduplicated by `file:line`; a per-crate loop
double-counts, see #572). This clears the 217 that clippy can rewrite
itself, leaving 127 that need a human decision.

Applied with `cargo clippy --fix`, then reviewed as a diff and followed
by `cargo fmt --all`. Nothing is suppressed: no `#[allow]` and no
`#[expect]` is added anywhere in this change. The largest classes were
`collapsible_if` (127), `unused_imports` (42), `doc_lazy_continuation`
and `doc_overindented_list_items` (partial), `needless_range_loop`
(partial), `map_clone`, `clone_on_copy`, `question_mark`,
`unnecessary_cast`, `needless_borrows_for_generic_args`,
`derivable_impls`, `useless_conversion`, `explicit_auto_deref`,
`unnecessary_map_or`, `len_zero`, `bool_assert_comparison`,
`manual_is_multiple_of` and `manual_range_contains`.

One autofix was wrong and is corrected by hand.
`crates/signex-app/src/library/editor/symbol/active_bar/mod.rs` had
`use std::path::PathBuf` deleted because the lib compilation unit does
not use it — but the in-src `#[cfg(test)] mod tests` does, through
`use super::*`, so the test unit stopped compiling with
`error[E0433]: cannot find type PathBuf in this scope`. The import is
moved into the test module, which is where the only use lives. This is
the same per-compilation-unit blind spot documented for `#[expect]` in
#569; `--fix` has it too, so its output must be compiled for
`--all-targets`, not just reviewed.

`cargo test --workspace` passes with 0 failures. The god-file,
lint-lockstep and discarded-Task gates all pass.

Part of #539.

+++ crates/signex-app/src/app/view/overlays/mod.rs
@@ -23,7 +23,7 @@ impl Signex {
     /// the dwell timer was running).
     pub(super) fn view_hover_tooltip(&self) -> Option<Element<'_, Message>> {
         use iced::widget::{column, container, text};
-        use iced::{Background, Border, Color};
+        use iced::{Background, Border};
 
         let interaction = &self.interaction_state;
         let uuid = interaction.hover_symbol_uuid?;
@@ -370,7 +370,7 @@ impl Signex {
         // CONTEXT_MENU_WIDTH; flip-up / flip-left when the click
         // lands too close to an edge.
         let (win_w, win_h) = self.ui_state.window_size;
-        let menu_w = Self::CONTEXT_MENU_WIDTH as f32;
+        let menu_w = Self::CONTEXT_MENU_WIDTH;
         let est_menu_h: f32 = 22.0 * 22.0 + 8.0;
         let edge_margin: f32 = 4.0;
         let x = if ctx_menu.x + menu_w + edge_margin > win_w {
@@ -470,7 +470,7 @@ impl Signex {
         let menu = self.view_tab_context_menu(tab_ctx);
         // Conservative footprint matches the project-tree menu so
         // the two visually align.
-        let menu_w = Self::CONTEXT_MENU_WIDTH as f32;
+        let menu_w = Self::CONTEXT_MENU_WIDTH;
         let est_menu_h: f32 = 5.0 * 22.0 + 8.0;
         let (win_w, win_h) = ui.window_size;
         let edge_margin: f32 = 4.0;
@@ -513,7 +513,7 @@ impl Signex {
         // Conservative footprint: at most 6 rows × 22 px + 8 px
         // padding. Width matches the canvas menu so the two look
         // consistent.
-        let menu_w = Self::CONTEXT_MENU_WIDTH as f32;
+        let menu_w = Self::CONTEXT_MENU_WIDTH;
         let est_menu_h: f32 = 6.0 * 22.0 + 8.0;
         let (win_w, win_h) = ui.window_size;
         let edge_margin: f32 = 4.0;
